requirements.txt
asgiref==3.8.1
Django==4.2
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
redis==5.0.8
sqlparse==0.5.1
tzdata==2024.1
